Version 0.4. Reworked sessions.
Major refactoring of sessions. An httpkom session is now more like a
LysKOM session, which is just a connection to the LysKOM server. You
need to first create a session before you can login, but you can also
logout and login in again without having to re-create the
session. This is a first step towards supporting multiple servers and
multiple sessions (not requiring cookies).