Fix Flask app initialization to work correctly by requiring explicit init.
When httpkom.main was added the initialization order was broken, as calling
httpkom.main would always import httpkom.__init__ and initialize
the Flask app object. The function httpkom.main.run_http_server() assumed
it was possible to modify the os.environ HTTPKOM_SETTINGS variable and
get the expected effect, but it actually didn't have any effect as the app
object was already initialized.